Go version developed under SVN
This is intended to work the same as the PHP version--and, in fact, the simulation works quite nicely, and substantially faster than the PHP version. We have not yet implemented reading parameters from the INI file, because there is no simple equivalent in go (that I can find) to PHP's parse_ini_file(). I have searched the internet, and while there seem to be other user-created libraries available, I'm thinking I may as well write my own. After all, the goal *is* to get some practice writing Go code.